# specs/001-uc-spec-union/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CMS API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /schedule/{scheduleId}/publish:
    post:
      summary: Publish final conference schedule
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /referee/invitations/{invitationId}/response:
    post:
      summary: Accept or reject review invitation
      parameters:
        - in: path
          name: invitationId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /registrations/{registrationId}/confirmation:
    get:
      summary: Receive payment confirmation ticket
      parameters:
        - in: path
          name: registrationId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK